make xds dump available in a non-printing function
